test(order-resolution): テストファイルを分割
